Ship/player faction(s), relation marker in OV
Faction for ships, as well as faction for player. Unable to select other
factions' ships. Relation marker in overview (not yet implemented mix,
not yet in game anyways). Redesigned the default UI and removed the
fleet overview and hostiles overview (as that is easily readable from
the standard overview, + that will be more easily filtered later).